Update steganography.py
Please note that I understand that the variable names may be vague, unhelpful or irrelevant as well as poor conventions. The reason for this is that this entire steganography problem was tackled by building it as very small, manageable chunks of code. Therefore descriptive variable names didn't make much sense as everything I built in small parts was virtually useless by themselves. By the time I put everything together into a class I just wanted to make sure it was finished and worked rather than to fiddle with conventions.